+++ src/Serwisant/SerwisantApi/AccessToken.php
@@ -6,9 +6,6 @@ interface AccessToken
 {
   const HOST = 'serwisant.online';
 
-  const URL = 'https://serwisant.online/oauth/token';
-  const REVOKE_URL = 'https://serwisant.online/oauth/revoke';
-
   public function get();
 
   public function refresh();